packages/renderer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "figma_component_dsl"
version = "0.1.0"
description = "PyCairo-based renderer for Figma Component DSL"
requires-python = ">=3.10"
dependencies = [
    "pycairo>=1.27.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
figma-dsl-render = "figma_component_dsl.renderer:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
